Restructured doc-browser.py
